First commit
Everything is done except for textures and normal maps. Controlls are available in editor with [shift + left mouse] - to add a point and [shift + right mouse] - to delete a point. Adding and deleting points work correctly only in ISO(isolated mode of camera in Scene View) with Z-axis maid ortigonal to a viewer. However while moving added points it doesnt matter how camera is rotated.